+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 3.7-beta
 -------------------
 - RTL layout issues #214 in rtl.css
+- INSERT INTO case to Replace empty strings ('') with NULL values in database.inc.php
 
 Changes in 3.6.1
 ----------------